+++ src/common/config/CodeOutlineConfig.ts
@@ -1,0 +1,15 @@
+/**
+ * Configuration interface for the code outline feature.
+ * 
+ * @export
+ * @interface CodeOutlineConfig
+ */
+export interface CodeOutlineConfig {
+    /**
+     * Defined if the code outline feature is enabled or not.
+     * 
+     * @type {boolean}
+     * @memberof CodeOutlineConfig
+     */
+    outlineEnabled: boolean;
+}