+++ src/extension/VscodeExtensionConfig.ts
@@ -2,6 +2,7 @@ import { injectable } from 'inversify';
 import { workspace } from 'vscode';
 
 import { ExtensionConfig, ResolverConfig } from '../common/config';
+import { CodeOutlineConfig } from '../common/config/CodeOutlineConfig';
 import { GenerationOptions } from '../common/ts-generation';
 import { ImportGroup, ImportGroupSetting, ImportGroupSettingParser, RemainImportGroup } from './import-grouping';
 
@@ -17,6 +18,7 @@ const sectionKey = 'typescriptHero';
 @injectable()
 export class VscodeExtensionConfig implements ExtensionConfig {
     private resolverConfig: ResolverConfig = new VscodeResolverConfig();
+    private codeOutlineConfig: CodeOutlineConfig = new VscodeCodeOutlineConfig();
 
     /**
      * The actual log level.
@@ -39,6 +41,17 @@ export class VscodeExtensionConfig implements ExtensionConfig {
     public get resolver(): ResolverConfig {
         return this.resolverConfig;
     }
+
+    /**
+     * Configuration object for the code outline extension.
+     * 
+     * @readonly
+     * @type {CodeOutlineConfig}
+     * @memberof VscodeExtensionConfig
+     */
+    public get codeOutline(): CodeOutlineConfig {
+        return this.codeOutlineConfig;
+    }
 }
 
 /**
@@ -56,7 +69,8 @@ class VscodeResolverConfig implements ResolverConfig {
      * @memberof VscodeResolverConfig
      */
     public get insertSpaceBeforeAndAfterImportBraces(): boolean {
-        return workspace.getConfiguration(sectionKey).get<boolean>('resolver.insertSpaceBeforeAndAfterImportBraces') || true;
+        const value = workspace.getConfiguration(sectionKey).get<boolean>('resolver.insertSpaceBeforeAndAfterImportBraces');
+        return value !== undefined ? value : true;
     }
 
     /**
@@ -68,7 +82,8 @@ class VscodeResolverConfig implements ResolverConfig {
      * @memberof VscodeResolverConfig
      */
     public get insertSemicolons(): boolean {
-        return workspace.getConfiguration(sectionKey).get<boolean>('resolver.insertSemicolons') || true;
+        const value = workspace.getConfiguration(sectionKey).get<boolean>('resolver.insertSemicolons');
+        return value !== undefined ? value : true;
     }
 
     /**
@@ -123,7 +138,8 @@ class VscodeResolverConfig implements ResolverConfig {
      * } from 'whatever';
      */
     public get multiLineTrailingComma(): boolean {
-        return workspace.getConfiguration(sectionKey).get<boolean>('resolver.multiLineTrailingComma') || true;
+        const value = workspace.getConfiguration(sectionKey).get<boolean>('resolver.multiLineTrailingComma');
+        return value !== undefined ? value : true;
     }
 
     /**
@@ -134,7 +150,8 @@ class VscodeResolverConfig implements ResolverConfig {
      * @memberof ResolverConfig
      */
     public get disableImportSorting(): boolean {
-        return workspace.getConfiguration(sectionKey).get<boolean>('resolver.disableImportsSorting') || false;
+        const value = workspace.getConfiguration(sectionKey).get<boolean>('resolver.disableImportsSorting');
+        return value !== undefined ? value : false;
     }
 
     /**
@@ -190,5 +207,26 @@ class VscodeResolverConfig implements ResolverConfig {
             stringQuoteStyle: this.stringQuoteStyle,
             tabSize: this.tabSize,
         };
+    }
+}
+
+/**
+ * Configuration interface for the code outline feature.
+ * 
+ * @class VscodeCodeOutlineConfig
+ * @implements {CodeOutlineConfig}
+ */
+class VscodeCodeOutlineConfig implements CodeOutlineConfig {
+
+    /**
+     * Defined if the code outline feature is enabled or not.
+     * 
+     * @readonly
+     * @type {boolean}
+     * @memberof VscodeCodeOutlineConfig
+     */
+    public get outlineEnabled(): boolean {
+        const value = workspace.getConfiguration(sectionKey).get<boolean>('codeOutline.enabled');
+        return value !== undefined ? value : true;
     }
 }