+++ CHANGELOG.md
@@ -8,7 +8,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 #### Fixed
 - Boolean settings are returned correctly now
-- Code outline does correctly jump to the selected element
+- Code outline does correctly jump to the selected element ([#219](https://github.com/buehler/typescript-hero/issues/219))
 
 ## [1.0.0] :shipit:
 #### Added