+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+
+## [1.1.0]
 #### Added
 - View (in the explorer) with the code outline (code structure) of the actual file - Jumps to code on click
 
@@ -196,7 +198,8 @@ The whole refactoring is part of ([#143](https://github.com/buehler/typescript-h
 - Various bugs in AST parsing
 
 
-[Unreleased]: https://github.com/buehler/typescript-hero/compare/v1.0.0...master
+[Unreleased]: https://github.com/buehler/typescript-hero/compare/v1.1.0...master
+[1.1.0]: https://github.com/buehler/typescript-hero/compare/v1.0.0...v1.1.0
 [1.0.0]: https://github.com/buehler/typescript-hero/compare/v0.13.2...v1.0.0
 [0.13.2]: https://github.com/buehler/typescript-hero/compare/v0.13.1...v0.13.2
 [0.13.1]: https://github.com/buehler/typescript-hero/compare/v0.13.0...v0.13.1